Move Plantmaster into interface folder
